Refactor DashboardContent component to replace ResizableBox with DraggableResizableBox, simplifying the code and enhancing drag-and-drop functionality. Removed unused drag logic and styles for improved maintainability.
