+++ src/app/features/dashboards/components/DashboardContent.tsx
@@ -3,22 +3,13 @@ import { Add as AddIcon } from '@mui/icons-material';
 import { useDashboard } from '../context/DashboardContext';
 import { useState } from 'react';
 import { CreateDashboardModal } from './CreateDashboardModal';
-import { useDrag, useDrop } from 'react-dnd';
-import { ResizableBox } from 'react-resizable';
-import 'react-resizable/css/styles.css';
+import { DraggableResizableBox } from './DraggableResizableBox';
+import { useDrop } from 'react-dnd';
 
 export const DashboardContent = () => {
   const { dashboards, selectedDashboard, setSelectedDashboard, createDashboard } = useDashboard();
   const [isCreateModalOpen, setIsCreateModalOpen] = useState(false);
   const [position, setPosition] = useState({ x: 0, y: 0 });
-  const [size, setSize] = useState({ width: 300, height: 200 });
-
-  const [{ isDragging }, drag] = useDrag(() => ({
-    type: 'BOX',
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
-  }));
 
   const [, drop] = useDrop(() => ({
     accept: 'BOX',
@@ -32,6 +23,10 @@ export const DashboardContent = () => {
       }
     },
   }));
+
+  const dropRef = (node: HTMLDivElement | null) => {
+    drop(node);
+  };
 
   if (!selectedDashboard) {
     return (
@@ -133,67 +128,13 @@ export const DashboardContent = () => {
     );
   }
 
-  const dropRef = (node: HTMLDivElement | null) => {
-    drop(node);
-  };
-
-  const dragRef = (node: HTMLDivElement | null) => {
-    drag(node);
-  };
-
   return (
     <div ref={dropRef} style={{ minHeight: '100vh', position: 'relative', backgroundColor: '#18181b', padding: 24 }}>
-      <ResizableBox
-        width={size.width}
-        height={size.height}
-        onResizeStop={(e, { size: newSize }) => setSize(newSize)}
-        minConstraints={[200, 150]}
-        maxConstraints={[800, 600]}
-        resizeHandles={['se']}
-        style={{
-          position: 'absolute',
-          left: position.x,
-          top: position.y,
-          backgroundColor: 'rgba(255, 255, 255, 0.05)',
-          borderRadius: 4,
-          padding: 16,
-          border: '1px solid rgba(255, 255, 255, 0.1)',
-          cursor: 'move',
-          opacity: isDragging ? 0.5 : 1,
-        }}
-      >
-        <div ref={dragRef} style={{ height: '100%', width: '100%' }}>
-          <Typography variant="h4" sx={{ color: '#e5e7eb', mb: 2 }}>
-            {selectedDashboard.title}
-          </Typography>
-          <Typography sx={{ color: '#6b7280' }}>
-            {selectedDashboard.description}
-          </Typography>
-        </div>
-      </ResizableBox>
-
-      <style jsx global>{`
-        .react-resizable {
-          padding: 0 !important;
-        }
-
-        .react-resizable-handle {
-          background-color: rgba(255, 255, 255, 0.1);
-          border: 1px solid rgba(255, 255, 255, 0.2);
-          border-radius: 2px;
-          padding: 0;
-          width: 10px;
-          height: 10px;
-        }
-        .react-resizable-handle:hover {
-          background-color: rgba(255, 255, 255, 0.2);
-        }
-        .react-resizable-handle-se {
-          bottom: 0;
-          right: 0;
-          cursor: se-resize;
-        }
-      `}</style>
+      <DraggableResizableBox
+        title={selectedDashboard.title}
+        description={selectedDashboard.description}
+        position={position}
+      />
     </div>
   );
 };
